+++ pkgs/dart_mcp_server/CHANGELOG.md
@@ -1,4 +1,10 @@
-# 0.1.2 (Dart SDK 3.11.0) - WIP
+# 0.1.3 (Dart SDK 3.12.0) - WIP
+
+- Add additional analytics for initialization events, and various list* method
+  calls. This will help understand how the server is being used by different
+  clients, and what features clients support.
+
+# 0.1.2 (Dart SDK 3.11.0)
 
 - Add `--tools=dart|all` argument to allow enabling only vanilla Dart tools for
   non-flutter projects.